Create TestBuild.txt
These are the commands to build and test the Source functionality using Fortran as the main program and calling the C++ objects.  This fills the C++ Source object and Samples it returning the Spectral sampling results.  

This is a test for using this C++ code with source.f in FLUKA